feat: 카페별 동적 OG 메타태그 구현
- page.tsx를 Server Component로 전환하고 generateMetadata 추가
- cafeId 쿼리 파라미터로 카페 데이터 조회 후 카페별 title/description/image OG 생성
- 카페 미선택 또는 데이터 없을 시 사이트 기본 OG로 fallback
- 클라이언트 로직은 HomeClient.tsx로 분리 (useSearchParams Suspense 경계 준수)
- description: 주소·이용나이·운영요일 레이블 포함
